+++ persistence/mongo/pom.xml
@@ -7,8 +7,8 @@
         <version>1.2.0-SNAPSHOT</version>
         <relativePath>../../pom.xml</relativePath>
     </parent>
-    <artifactId>persistence-mongodb</artifactId>
-    <name>persistence-mongodb</name>
+    <artifactId>persistence-mongo</artifactId>
+    <name>persistence-mongo</name>
     <properties>
         <root.basedir>${project.parent.basedir}</root.basedir>
     </properties>
