+++ persistence/mongo/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/persistence/mongo/PersistenceMongo.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.fraunhofer.iosb.ilt.faaast.service.persistence.mongodb;
+package de.fraunhofer.iosb.ilt.faaast.service.persistence.mongo;
 
 import com.mongodb.ConnectionString;
 import com.mongodb.MongoClientSettings;
