+++ persistence/mongo/src/test/java/de/fraunhofer/iosb/ilt/faaast/service/persistence/mongo/PersistenceMongoTest.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.fraunhofer.iosb.ilt.faaast.service.persistence.mongodb;
+package de.fraunhofer.iosb.ilt.faaast.service.persistence.mongo;
 
 import de.flapdoodle.embed.mongo.distribution.Version;
 import de.flapdoodle.embed.mongo.transitions.Mongod;
