+++ starter/pom.xml
@@ -90,7 +90,7 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>persistence-mongodb</artifactId>
+            <artifactId>persistence-mongo</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
